requirements.txt
aiohttp==3.13.5
aiosqlite==0.22.1
anthropic==0.92.0
beautifulsoup4==4.14.3
chromadb==1.5.7
fastapi==0.135.3
google-api-python-client==2.194.0
google-auth-oauthlib==1.3.1
google-auth==2.49.1
groq==1.1.2
httpx==0.28.1
langgraph-supervisor==0.0.31
langgraph==1.1.6
mem0ai==1.0.11
numpy==2.4.4
ollama==0.6.1
onnxruntime==1.24.4
openwakeword==0.6.0
pydantic==2.12.5
python-dotenv==1.2.2
requests==2.33.1
semantic-router==0.0.20
sentence-transformers==5.3.0
sounddevice==0.5.5
soundfile==0.13.1
textual==8.2.3
uvicorn==0.44.0
